gateway: harden relay task teardown
Reap completed rendezvous tasks during steady state and exercise listener cleanup through the production control loop and a real local socket session.
